+++ lib/dradis-appspider.rb
@@ -2,7 +2,7 @@
 require 'dradis-plugins'
 
 # load this add-on's engine
-require 'dradis/plugins/ntospider'
+require 'dradis/plugins/appspider'
 
 # load supporting NTOSpider classes
-require 'ntospider/vuln'
+require 'appspider/vuln'