+++ lib/dradis/plugins/appspider/engine.rb
@@ -1,9 +1,9 @@
-module Dradis::Plugins::NTOSpider
+module Dradis::Plugins::AppSpider
   class Engine < ::Rails::Engine
     isolate_namespace Dradis::Plugins::NTOSpider
 
     include ::Dradis::Plugins::Base
-    description 'Processes NTOSpider reports'
+    description 'Processes AppSpider reports'
     provides :upload
   end
 end